исправлены ошибки верстки, исправлены презенторы. Доработан мд файл убраны слишком длинные строки. Некоторые изображения заменены на символы
