+++ FakeNFT/Scenes /Profile/ProfilePresenter.swift
@@ -4,19 +4,14 @@
 //
 //  Created by Владимир Клевцов on 15.1.24..
 //
-import UIKit
+import Foundation
 
 final class ProfilePresenter {
     private let profileService = ProfileService.shared
     weak var viewController: ProfileViewController?
-    
+
     init(viewController: ProfileViewController) {
         self.viewController = viewController
-    }
-    func presentWeb(gesture: UITapGestureRecognizer) {
-        if let urlString = profileService.website, let url = URL(string: "https://" + urlString) {
-            UIApplication.shared.open(url, options: [:], completionHandler: nil)
-        }
     }
     func presentEdit() {
             let editProfileViewController = EditProfileViewController()
