+++ FakeNFT/Services/Requests/RequestConstants.swift
@@ -1,4 +1,4 @@
 enum RequestConstants {
-    #warning("insert your baseUrl <id>.mockapi.io")
+    // #warning("insert your baseUrl <id>.mockapi.io")
     static let baseURL = "e2ac207e-630b-486b-9153-b5d079c4bcee.mockapi.io"
 }
