+++ decomposition/ProfileDecom.md
@@ -2,10 +2,10 @@
 Способ верстки: код
 сслыка на доску https://github.com/users/vovameister/projects/5/views/1
 1/3
-верстка основного экрана профиля 2 часа
-вертска редактирования профиля 2 часа
-добавление функционала создания аватара 1 час
-сохранение информации из профиля в UserDefolt 1 час
+верстка основного экрана профиля 2 часа (затрачено 2 часа)
+вертска редактирования профиля 2 часа (завтрачено 2 часа)
+добавление функционала создания аватара 1 час (затрачено 2 часа)
+сохранение информации из профиля в UserDefolt 1 час (затрачено 3 часа)
 2/3
 верстка экрана мои нфт 2 часа 
 добавление фильтров 1-2 часа
